Refactored calculateDamage in enemy to generateDamage for more consistent names Removed calculate damage from gameCharacter Added using items functionality to player Modified consumable items and item generation Added turn constructor for using an item Made enemies attack and kill you

+++ battlehauz/src/models/items/ConsumeableItem.java
@@ -6,6 +6,6 @@ public abstract class ConsumeableItem extends Item {
         super(iName, iBuyingPrice, iSellingPrice);
     }
 
-    public abstract int useItem();
+    public abstract double useItem();
 
 }
